Allows analytics to initialize after consent is given.
