+++ lib/hooks/useAnalytics.js
@@ -1,14 +1,76 @@
 import ReactGA from 'react-ga';
 import storage from 'local-storage-fallback';
+import { getCookieConsentValue } from 'react-cookie-consent';
 import { getQueryVariable } from '../utils.js';
+
+export const trackReadingHistoryWithPageView = (hookObj) => {
+  hookObj.logReadingHistory();
+  const readingHistory = hookObj.summarizeReadingHistory();
+  hookObj.setDimension('dimension2', readingHistory);
+  return {
+    dimension2: readingHistory,
+  };
+};
+
+export const initialize = (hookObj) => {
+  const {
+    init,
+    trackPageViewedWithDimensions,
+    setDimension,
+    logReadingHistory,
+    summarizeReadingHistory,
+    donorStatusFromStorage,
+    newsletterStatusFromStorage,
+    trackMailChimpParams,
+  } = hookObj;
+
+  function trackNewsletterVisits(trackMailChimpParams) {
+    let hitFromSubscriber = trackMailChimpParams();
+    let isLoggedSubscriber = newsletterStatusFromStorage();
+    if (hitFromSubscriber || isLoggedSubscriber) {
+      setDimension('dimension5', true);
+      return {
+        dimension5: true,
+      };
+    } else {
+      return {};
+    }
+  }
+
+  init(process.env.NEXT_PUBLIC_GA_TRACKING_ID);
+  let readingDimensionsData = trackReadingHistoryWithPageView(hookObj);
+  let newsletterDimensionsData = trackNewsletterVisits(trackMailChimpParams);
+
+  let dimensionsData = {
+    ...readingDimensionsData,
+    ...newsletterDimensionsData,
+  };
+
+  let donorStatus = donorStatusFromStorage();
+  if (donorStatus) {
+    setDimension('dimension4', true);
+    dimensionsData['dimension4'] = true;
+  }
+
+  let pagePath = window.location.pathname + window.location.search;
+  if (!/tinycms/.test(pagePath)) {
+    trackPageViewedWithDimensions(pagePath, dimensionsData);
+  }
+};
 
 export const useAnalytics = () => {
   return {
     init: (trackingId) => {
-      console.log('Initialising ReactGA with trackingId:', trackingId);
-      ReactGA.initialize(trackingId, { debug: false });
+      if (getCookieConsentValue()) {
+        console.log('Initialising ReactGA with trackingId:', trackingId);
+        ReactGA.initialize(trackingId, { debug: true });
+      }
     },
     trackPageViewed: (path) => {
+      if (!getCookieConsentValue()) {
+        return true;
+      }
+
       if (path) {
         return ReactGA.pageview(path);
       }
@@ -17,14 +79,26 @@ export const useAnalytics = () => {
       );
     },
     trackEvent: (params) => {
+      if (!getCookieConsentValue()) {
+        return true;
+      }
+
       ReactGA.event(params);
     },
     setDimension: (dimension, value) => {
+      if (!getCookieConsentValue()) {
+        return true;
+      }
+
       let fields = {};
       fields[dimension] = value;
       ReactGA.set(fields);
     },
     trackPageViewedWithDimensions: (path, dimensions) => {
+      if (!getCookieConsentValue()) {
+        return true;
+      }
+
       // let fields = {};
       // fields[dimensionName] = dimensionData;
 
@@ -37,6 +111,10 @@ export const useAnalytics = () => {
       );
     },
     trackMailChimpParams: () => {
+      if (!getCookieConsentValue()) {
+        return true;
+      }
+
       let campaignId = getQueryVariable('mc_cid');
       let emailId = getQueryVariable('mc_eid');
 
@@ -52,9 +130,17 @@ export const useAnalytics = () => {
       }
     },
     newsletterStatusFromStorage: () => {
+      if (!getCookieConsentValue()) {
+        return true;
+      }
+
       return storage.getItem('TNCNewsletterSubscriber');
     },
     logReadingHistory: () => {
+      if (!getCookieConsentValue()) {
+        return true;
+      }
+
       let history = JSON.parse(storage.getItem('TNCHistory'));
       if (!history || !Array.isArray(history)) {
         console.log('logReadingHistory: resetting history!', history);
@@ -74,9 +160,17 @@ export const useAnalytics = () => {
       storage.setItem('TNCHistory', JSON.stringify(history));
     },
     donorStatusFromStorage: () => {
+      if (!getCookieConsentValue()) {
+        return true;
+      }
+
       return storage.getItem('TNCDonor');
     },
     checkReferrer: (referrer) => {
+      if (!getCookieConsentValue()) {
+        return true;
+      }
+
       if (!referrer) {
         return false;
       }
@@ -89,6 +183,10 @@ export const useAnalytics = () => {
       return false;
     },
     summarizeReadingHistory: () => {
+      if (!getCookieConsentValue()) {
+        return true;
+      }
+
       let contactFrequency = '0 posts';
       const contactHistory = JSON.parse(storage.getItem('TNCHistory'));
